rename: Theme -> main 폴더명 변경
