+++ src/components/common/main/ChooseFriend.css
