+++ src/components/common/main/GiftRanking.css
