+++ src/components/common/main/RankingBoard.css
