+++ src/components/pages/theme/Theme.js
@@ -1,0 +1,12 @@
+import React from 'react'
+
+import Header from '@/components/Header/Header'
+
+
+export default function Theme() {
+  return (
+    <div>
+      <Header/>
+    </div>
+  )
+}
